Update welcome.yml
Added GitHub CLI (gh) -> add comment when pull request opened. 
GITHUB_TOKEN environment variable to the value of the GITHUB_TOKEN secret = allow GitHub CLI to post comment. Which is an installation access token, created when the workflow runs. 
PR_URL set environment variable to the URL of the newly created pull request, and we use this in the gh command.